✨ Allow mod commands to be run via chat
- Patched the ChatBox::runCommand() to execute the registered custom
  command if the command name matches and stop the further execution of
  the method.
- Each command's method is now retreived as a delegate instead of
  directly adding then to `ConsoleCommands` property like before.
- This allows them to simultaneously be added to `ConsoleCommands` and a
  `Commands` dictionary which is used in the `RunCommandPatch`.

Mainly added this feature as the Chat Commands mod no longer works and
the building commands are better used in the chat box rather than the
terminal. (#309)
Also, I will be adding a command to repeat the last spoken texts by the
screen reader and that will also not be very intuitive with commands not
being able to run in chat.