✨ Patch Home Renovations Menu
Use the PrimaryInfoKey in the menu to snap the mouse to the renovation
area and speak the details about the affected area.
Then use LeftClick to renovate.

The menu is similar to the building menu but as there is no use of
wandering aroung the map except to select the specified area, this
work around should be enough

closes #220