feat: Remove name and label as fallback to ScreenReaderText
Previously, if a menu is not defined as manually patched i.e., the auto-patch logic (IClickableMenuPatch::DrawPatch and IClickableMenuPatch::DrawHoverTextPatch) is not aborted, the following things happen...
Case 1: The DrawPatch checks for the hovered element and it doesn't find one so it lets the hover patch to execute and speaks the hovered text if any.
Case 2: The draw patch checks for the hovered element and finds one so it speaks its ScreenReaderText property, and if that is unset then it fallbacks to speak either the name or label property which is present in all the elements. And then it aborts the hover patch.

Well, the fallback thing can cause an issue if for example a mod author does not set ScreenReaderText (which will be the case for all mods as this thing is new) and sets an arbitrary value in name or label (I've seen these properties used as kinda like ids in a few places in vanilla code, maybe some mod authors copy this paradigm) but, they correctly set and display the hover text for the hovered element. And because the current logic takes the fallback properties into account (name and label), it will abort speaking hover text and instead speak the arbitrary value.

This is what's happening with spacecore's skills page, the hover texts are displayed correctly but since the logic aborts speaking them, daniel reported the mod speaking numbers instead.
So I'm removing the fallbacks, imo it will only complicate things further. The downside is that some mods might've already been using name and label correctly but that won't be spoken.